cast cycles to unsigned for rt_logf

+++ examples/cycle/queue.c
@@ -35,5 +35,5 @@ int main(void)
     rt_cycle_enable();
     rt_start();
 
-    rt_logf("cycles = %u\n", cycles);
+    rt_logf("cycles = %u\n", (unsigned)cycles);
 }
